started adding options page
