+++ src/client/object/Options/OptionItem/OptionItemEdit.js
@@ -1,0 +1,40 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { sortOptions } from '../../../../common/helpers/wObjectHelper';
+
+const OptionItemEdit = ({ option, wobject }) => (
+  <div className="Options__block">
+    {' '}
+    {option[1].some(el => el.author_permlink === wobject.author_permlink) && (
+      <div className="Options__option-category">{option[0]}: </div>
+    )}
+    {option[1]
+      ?.sort((a, b) => sortOptions(a, b))
+      .map(
+        el =>
+          el.author_permlink === wobject.author_permlink && (
+            <div key={el.author_permlink}>
+              {el.body.position}
+              {el.body.position ? '.' : ''} {el.body.value}{' '}
+              {el.body.image && (
+                <div>
+                  <img
+                    className="Options__my-pictures"
+                    src={el.body.image}
+                    alt="option"
+                    key={el.permlink}
+                  />
+                </div>
+              )}
+            </div>
+          ),
+      )}
+  </div>
+);
+
+OptionItemEdit.propTypes = {
+  wobject: PropTypes.shape().isRequired,
+  option: PropTypes.arrayOf().isRequired,
+};
+
+export default OptionItemEdit;
