+++ src/client/object/Options/OptionsPage.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const OptionsPage = () => <div>lallakjshssfhjkkl</div>;
+
+export default OptionsPage;
