+++ src/routes/components.js
@@ -129,6 +129,7 @@ export default {
   FormPage: loableComponent(import('../client/object/FormPage/FormPage')),
   WidgetPage: loableComponent(import('../client/object/WidgetPage/WidgetPage')),
   DescriptionPage: loableComponent(import('../client/object/Description/DescriptionPage')),
+  OptionsPage: loableComponent(import('../client/object/Options/OptionsPage')),
   DepartmentsPage: loableComponent(import('../client/object/DepartmentsPage/DepartmentsPage')),
   WebsiteSignIn: loableComponent(import(`../client/websites/WebsiteSignIn/WebsiteSignIn`)),
   VipTicketsSetting: loableComponent(import(`../client/settings/Viptickets/VipTicketsSetting`)),
