+++ src/routes/configs/routes.js
@@ -424,6 +424,11 @@ const routes = {
           component: Views.DescriptionPage,
         },
         {
+          path: '/options',
+          exact: true,
+          component: Views.OptionsPage,
+        },
+        {
           path: '/departments/:permlink',
           exact: true,
           component: Views.DepartmentsPage,
